gdc-hardware-api/requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
flask
flask-cors
